requirements.txt
Flask==3.0.0
Pillow==10.0.0
opencv-python==4.8.1.78
numpy==1.24.3
pytesseract==0.3.10
pyttsx3==2.90
googletrans==3.1.0a0
torch==2.1.0
torchvision==0.16.0
rembg==2.0.50
transformers==4.35.0
pdf2image==1.16.3
fpdf==1.7.2
python-dotenv==1.0.0
Werkzeug==3.0.1 